services/service_a/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3

prometheus-client==0.20.0

opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-requests==0.47b0
